// src/types/patients.ts
export type Patients = {
  id?: number;
  first_name?: string;
  last_name?: string;
  gender?: string;
  date_of_birth?: string;
  phone_number?: string;
  oib?: string;
  address?: string;
  city?: string;
  country?: string;
  email?: string;
  mbo?: string;
  newsletter?: boolean;
  created_at?: string;
};